another trick to force gpicker to popup on top
The problem that's addressed is that netbeans/emacs often triggers
gpicker by keypress and it's possible that netbeans will will receive
key release event after gpicker is started but before the window is
displayed. This seems to be quite typical on KDE4 with window
compositing turned on. In such case _WM_TIME of gpicker will be older
than _WM_TIME of netbeans causing window manager to keep netbeans
window above gpicker's.